x/fantasfive/types/message_create_team_test.go
package types

import (
	"testing"

	"fantasfive/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateTeam_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateTeam
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateTeam{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateTeam{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
